+++ test/dom_builder_template_test.dart
@@ -6,8 +6,11 @@ void main() {
     setUp(() {});
 
     test('parse: var 1', () {
-      var template = DOMTemplate.parse('My name is {{name}}!');
+      var source = 'My name is {{name}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Joe!'));
@@ -17,8 +20,11 @@ void main() {
     });
 
     test('parse: var 2', () {
-      var template = DOMTemplate.parse('My name is {{name}}!!!');
+      var source = 'My name is {{name}}!!!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Joe!!!'));
@@ -28,8 +34,11 @@ void main() {
     });
 
     test('parse: var 3', () {
-      var template = DOMTemplate.parse('My name is {{name}}');
+      var source = 'My name is {{name}}';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(2));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Joe'));
@@ -39,8 +48,11 @@ void main() {
     });
 
     test('parse: var 4', () {
-      var template = DOMTemplate.parse('My name is {{ name }}.');
+      var source = 'My name is {{ name }}.';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals('My name is {{name}}.'));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Joe.'));
@@ -50,8 +62,11 @@ void main() {
     });
 
     test('parse: var else', () {
-      var template = DOMTemplate.parse('name is {{?name}}unknown{{/}}!');
+      var source = 'name is {{?name}}unknown{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('name is Joe!'));
@@ -61,8 +76,11 @@ void main() {
     });
 
     test('parse: if 1', () {
-      var template = DOMTemplate.parse('name var {{:name}}present{{/}}!');
+      var source = 'name var {{:name}}present{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('name var present!'));
@@ -72,8 +90,11 @@ void main() {
     });
 
     test('parse: if 2', () {
-      var template = DOMTemplate.parse('name var {{:name}}present{{/name}}!');
+      var source = 'name var {{:name}}present{{/name}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals('name var {{:name}}present{{/}}!'));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('name var present!'));
@@ -83,8 +104,11 @@ void main() {
     });
 
     test('parse: not 1', () {
-      var template = DOMTemplate.parse('Name is {{!name}}empty{{/}}!');
+      var source = 'Name is {{!name}}empty{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({});
       expect(s1, equals('Name is empty!'));
@@ -94,8 +118,11 @@ void main() {
     });
 
     test('parse: not 2', () {
-      var template = DOMTemplate.parse('Name is {{!name}}empty{{/name}}!');
+      var source = 'Name is {{!name}}empty{{/name}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals('Name is {{!name}}empty{{/}}!'));
 
       var s1 = template.build({});
       expect(s1, equals('Name is empty!'));
@@ -105,8 +132,11 @@ void main() {
     });
 
     test('parse: if_else 1', () {
-      var template = DOMTemplate.parse('My name is {{:name}}Foo{{?}}Bar{{/}}!');
+      var source = 'My name is {{:name}}Foo{{?}}Bar{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Foo!'));
@@ -116,9 +146,11 @@ void main() {
     });
 
     test('parse: if_elseif_else 1', () {
-      var template = DOMTemplate.parse(
-          'My name is {{:name}}Foo{{?:surname}}Fii{{?}}Bar{{/}}!');
+      var source = 'My name is {{:name}}Foo{{?:surname}}Fii{{?}}Bar{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Foo!'));
@@ -131,9 +163,12 @@ void main() {
     });
 
     test('parse: if_elseif_elseif_else 1', () {
-      var template = DOMTemplate.parse(
-          'My name is {{:name}}Foo{{?:surname}}Fii{{?:nickname}}Fuu{{?}}Bar{{/}}!');
+      var source =
+          'My name is {{:name}}Foo{{?:surname}}Fii{{?:nickname}}Fuu{{?}}Bar{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Foo!'));
@@ -149,9 +184,11 @@ void main() {
     });
 
     test('parse: if_elseif_elseif 1', () {
-      var template =
-          DOMTemplate.parse('My name is {{:name}}Foo{{?:surname}}Fii{{/}}!');
+      var source = 'My name is {{:name}}Foo{{?:surname}}Fii{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'name': 'Joe'});
       expect(s1, equals('My name is Foo!'));
@@ -164,9 +201,11 @@ void main() {
     });
 
     test('parse: not_else 1', () {
-      var template =
-          DOMTemplate.parse('name is {{!name}}empty{{?}}present{{/}}!');
+      var source = 'name is {{!name}}empty{{?}}present{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({});
       expect(s1, equals('name is empty!'));
@@ -176,8 +215,11 @@ void main() {
     });
 
     test('parse: if{ var } 1', () {
-      var template = DOMTemplate.parse('E-mail is {{:name}}"{{email}}"{{/}}!');
+      var source = 'E-mail is {{:name}}"{{email}}"{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'email': 'a@b.c'});
       expect(s1, equals('E-mail is !'));
@@ -187,9 +229,11 @@ void main() {
     });
 
     test('parse: if{ var }_else 1', () {
-      var template = DOMTemplate.parse(
-          'E-mail is {{:name}}"{{email}}"{{?}}unknown_user{{/}}!');
+      var source = 'E-mail is {{:name}}"{{email}}"{{?}}unknown_user{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'email': 'a@b.c'});
       expect(s1, equals('E-mail is unknown_user!'));
@@ -199,9 +243,12 @@ void main() {
     });
 
     test('parse: if{ var_else }_else 1', () {
-      var template = DOMTemplate.parse(
-          'E-mail is {{:name}}"{{?email}}NULL_MAIL{{/}}"{{?}}unknown_user{{/}}!');
+      var source =
+          'E-mail is {{:name}}"{{?email}}NULL_MAIL{{/}}"{{?}}unknown_user{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'email': 'a@b.c'});
       expect(s1, equals('E-mail is unknown_user!'));
@@ -214,8 +261,11 @@ void main() {
     });
 
     test('parse: ifList 1', () {
-      var template = DOMTemplate.parse('Routes:{{*:routes}} [{{name}}]{{/}}!');
+      var source = 'Routes:{{*:routes}} [{{name}}]{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({
         'routes': [
@@ -233,8 +283,11 @@ void main() {
     });
 
     test('parse: ifList 2', () {
-      var template = DOMTemplate.parse('Route 1: {{routes.1.name}}!');
+      var source = 'Route 1: {{routes.1.name}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({
         'routes': [
@@ -259,8 +312,11 @@ void main() {
     });
 
     test('parse: ifList 3', () {
-      var template = DOMTemplate.parse('Routes:{{*:routes}} [{{.}}]{{/}}!');
+      var source = 'Routes:{{*:routes}} [{{.}}]{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({
         'routes': [
@@ -283,9 +339,11 @@ void main() {
     });
 
     test('parse: ifList_else 1', () {
-      var template = DOMTemplate.parse(
-          'Routes:{{*:routes}} [{{name}}]{{?}} NO ROUTES{{/}}!');
+      var source = 'Routes:{{*:routes}} [{{name}}]{{?}} NO ROUTES{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({
         'routes': [
@@ -303,8 +361,11 @@ void main() {
     });
 
     test('parse: notList 1', () {
-      var template = DOMTemplate.parse('Routes:{{!routes}}{{no_routes}}{{/}}!');
+      var source = 'Routes:{{!routes}}{{no_routes}}{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({
         'no_routes': 'Empty routes',
@@ -320,8 +381,11 @@ void main() {
     });
 
     test('parse: query element', () {
-      var template = DOMTemplate.parse('Element: {{:ok}}{{#element_x}}{{/}}!');
+      var source = 'Element: {{:ok}}{{#element_x}}{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var elementProvider = (q) {
         return q == '#element_x' ? 'XXX' : null;
@@ -335,9 +399,12 @@ void main() {
     });
 
     test('parse: ifEq{ var } 1', () {
-      var template = DOMTemplate.parse(
-          'Hello! Good {{:period=="am"}}morning{{?:period=="pm"}}afternoon{{?}}day{{/}}!');
+      var source =
+          'Hello! Good {{:period=="am"}}morning{{?:period=="pm"}}afternoon{{?}}day{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'period': 'am'});
       expect(s1, equals('Hello! Good morning!'));
@@ -350,9 +417,12 @@ void main() {
     });
 
     test('parse: ifEq{ var } 2', () {
-      var template = DOMTemplate.parse(
-          'Hello! Period: {{:period==hourPeriod}}match{{?}}no match{{/}}!');
+      var source =
+          'Hello! Period: {{:period==hourPeriod}}match{{?}}no match{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'period': 'am', 'hourPeriod': 'am'});
       expect(s1, equals('Hello! Period: match!'));
@@ -362,15 +432,52 @@ void main() {
     });
 
     test('parse: ifNotEq{ var } 2', () {
-      var template = DOMTemplate.parse(
-          'Hello! Period: {{:period!=hourPeriod}}diff{{?}}eq{{/}}!');
+      var source = 'Hello! Period: {{:period!=hourPeriod}}diff{{?}}eq{{/}}!';
+
+      var template = DOMTemplate.parse(source);
       expect(template.nodes.length, equals(3));
+      expect(template.toString(), equals(source));
 
       var s1 = template.build({'period': 'am', 'hourPeriod': 'am'});
       expect(s1, equals('Hello! Period: eq!'));
 
       var s2 = template.build({'period': 'am', 'hourPeriod': 'pm'});
       expect(s2, equals('Hello! Period: diff!'));
+    });
+
+    test('parse: ifList (menu)', () {
+      var source = 'Menu:\n'
+          '{{*:menu}}'
+          '- #{{route}}: {{name}}{{:current}} (active){{/}}\n'
+          '{{/}}';
+
+      var template = DOMTemplate.parse(source);
+      expect(template.nodes.length, equals(2));
+      expect(template.toString(), equals(source));
+
+      var s1 = template.build({
+        'menu': [
+          {'route': 'home', 'name': 'Welcome', 'current': true},
+          {'route': 'help', 'name': 'Help-me'},
+        ]
+      });
+      expect(
+          s1,
+          equals('Menu:\n'
+              '- #home: Welcome (active)\n'
+              '- #help: Help-me\n'));
+    });
+
+    test('tryParse', () {
+      var source1 = 'Hello! Period: {{:period!=hourPeriod}}diff{{?}}eq{{/}}!';
+      var source2 = 'Hello! Period: {{:period!=hourPeriod}}diff{{?}}eq!';
+
+      var template1 = DOMTemplate.tryParse(source1);
+      expect(template1.nodes.length, equals(3));
+      expect(template1.toString(), equals(source1));
+
+      var template2 = DOMTemplate.tryParse(source2);
+      expect(template2, isNull);
     });
   });
 }